+++ config/settings/base.py
@@ -68,6 +68,9 @@ LOCAL_APPS = [
     # Projectroles itself
     'projectroles.apps.ProjectrolesConfig',
 
+    # User Profile site app
+    'user_profile.apps.UserprofileConfig',
+
     # Example project app
     'example_project_app.apps.ExampleProjectAppConfig',