+++ user_profile/urls.py
@@ -2,10 +2,12 @@ from django.conf.urls import url
 
 from . import views
 
+app_name = 'user_profile'
+
 urlpatterns = [
     url(
-        regex=r'^(?P<username>[\w.@+-]+)$',
+        regex=r'^$',
         view=views.UserDetailView.as_view(),
-        name='user_detail'
+        name='detail'
     ),
 ]